// lib/jsonRpcEnv.d.ts
/**
 *
 */
export interface IMedMeJsonRpcEnv {
    CORE_API_ENDPOINT: string;
    CRAC_SLOTS_API_ENDPOINT: string;
    CRAC_API_ENDPOINT: string;
    CRAC3_API_ENDPOINT: string;
    OAUTH_OTP_SEND: string;
    OAUTH_OTP_VERIFY: string;
    OAUTH_OTP_WEBLOGIN: string;
    JSONRPC_REQUEST_DEBUG: boolean;
    OTP_REQUEST_DEBUG: boolean;
}